Clean up of Select Folder
I discussed my changes with DooM. My changes would make any directory with another directory named 'files' be valid.

DooM had suggestions to add boolean or three checks to verify if the folder is the SC folder. This was a good application of the code and I decided to apply it into this fork.